+++ src/constants/message.ts
@@ -1,0 +1,3 @@
+export const MESSAGE = {
+  cancel: '\n👋 Process was cancelled by user',
+};