+++ src/service/github-service/index.ts
@@ -12,6 +12,9 @@ export * from './extractRemoteIssueBranches';
 export * from './syncForkBranchAndUpdateLocal';
 export * from './switchBranch';
 export * from './deleteRemoteBranches';
+export * from './checkCurrentBranchIsIssueBranch';
+export * from './confirmToDeleteBranches';
+export * from './getBranchDiff';
 
 /** CLI */
 export * from './checkGithubAuth';