+++ src/service/github-service/pushToTargetBranch.ts
@@ -1,3 +1,4 @@
+import { ColorCode } from '@/constants/colors';
 import { DefaultConfig } from '@/types';
 import { executeCommand } from '@/utils/common-utils/executeCommand';
 
@@ -21,7 +22,12 @@ export function createPushToTargetBranchErrorMessage({
   REPO_OWNER?: string;
   REPO_NAME?: string;
 }) {
-  return `\n🚫 Failed to push to ${REPO_OWNER}/${REPO_NAME} repository. please check your env again`;
+  return `\n🚫 Failed to push to ${REPO_OWNER}/${REPO_NAME} repository.\n
+  ${ColorCode.yellow('hint: Updates were rejected because the tip of your current branch is behind')}
+  ${ColorCode.yellow('its remote counterpart. Integrate the remote changes (e.g.')}
+  ${ColorCode.yellow('git pull ...) before pushing again.')}
+  ${ColorCode.yellow('See the "Note about fast-forwards" in "git push --help" for details.')}
+  `;
 }
 
 export function createPushToTargetBranchSuccessMessage() {