+++ src/utils/common-utils/loadEnv.ts
@@ -1,9 +1,0 @@
-import dotenv from 'dotenv';
-import path from 'path';
-
-export function loadEnv() {
-  const environment = process.env.NODE_ENV;
-  const envFilePath = environment ? `.env.${environment}` : '.env';
-
-  dotenv.config({ path: path.resolve(process.cwd(), envFilePath) });
-}