+++ src/utils/pr-utils/inquirePRTitle.ts
@@ -5,9 +5,18 @@ interface InquirePRTitleParams {
 }
 
 export async function inquirePRTitle({ defaultValue }: InquirePRTitleParams = {}) {
+  const mergeText = 'Merge branch';
+
   const title = await input({
     message: 'Enter PR Title (default: latest commit title):',
-    default: defaultValue,
+    default: !defaultValue?.includes(mergeText) ? defaultValue : undefined,
+    validate: (value: string) => {
+      if (!value) {
+        return 'Please enter a PR title';
+      }
+
+      return true;
+    },
   });
 
   return title;